Remove the `serial` module entirely; move filesystem utilities into new `fs` module.
