+++ crates/notion-core/src/lib.rs
@@ -27,12 +27,12 @@ pub mod config;
 mod distro;
 pub mod env;
 mod event;
+pub(crate) mod fs;
 pub mod manifest;
 pub mod monitor;
 pub mod path;
 mod plugin;
 pub mod project;
-pub mod serial;
 pub mod session;
 pub mod shell;
 pub mod shim;
