+++ src/command/fetch.rs
@@ -1,6 +1,6 @@
 use semver::VersionReq;
 
-use notion_core::serial::version::parse_requirements;
+use notion_core::version::serial::parse_requirements;
 use notion_core::catalog::{parse_node_version, parse_yarn_version};
 use notion_core::session::{ActivityKind, Session};
 use notion_fail::{ExitCode, Fallible};
